starters/nlp-project-starter/nlp-project/requirements.txt
torch
datasets
torchmetrics==1.4.0
tqdm
pyyaml
numpy
